Fixed a major in the matrix memory management (shadow int to short conversion). Some other things.

+++ src/gcloadunit.cu
@@ -298,7 +298,7 @@ int GcLoadUnit::addVectorToDictMap(const std::vector<std::string> *vect) {
     return dc1;
 }
 
-void GcLoadUnit::appendMatrix(const unsigned short *mat1, unsigned short sizeofMat, unsigned short *gcMatrixData,
+void GcLoadUnit::appendMatrix(const unsigned short *mat1, unsigned long sizeofMat, unsigned short *gcMatrixData,
                               unsigned int *gcDictData, unsigned int *gcMatrixOffsets, unsigned int *gcMatrixSizes,
                               int *lastOffset,
                               int position) {
